TSConfig imports and reorganize
