+++ jest.config.cjs
@@ -14,11 +14,15 @@ module.exports = {
     },
   },
   moduleNameMapper: {
-    '^@canvas$': resolve(__dirname, './src/canvas'),
-    '^@canvas/(.*)$': resolve(__dirname, './src/canvas/$1'),
-    '^@ui$': resolve(__dirname, './src/ui'),
-    '^@ui/(.*)$': resolve(__dirname, './src/ui/$1'),
-    '^@db$': resolve(__dirname, './src/db'),
-    '^@db/(.*)$': resolve(__dirname, './src/db/$1'),
+    '^canvas$': resolve(__dirname, './src/canvas'),
+    '^canvas/(.*)$': resolve(__dirname, './src/canvas/$1'),
+    '^ui$': resolve(__dirname, './src/ui'),
+    '^ui/(.*)$': resolve(__dirname, './src/ui/$1'),
+    '^db$': resolve(__dirname, './src/db'),
+    '^db/(.*)$': resolve(__dirname, './src/db/$1'),
+    '^test-utils$': resolve(__dirname, './test-utils'),
+    '^test-utils/(.*)$': resolve(__dirname, './test-utils/$1'),
+    '^types$': resolve(__dirname, './src/types'),
+    '^types/(.*)$': resolve(__dirname, './src/types/$1'),
   },
 };
