+++ src/db/index.ts
@@ -2,5 +2,4 @@
 
 export * from './queries';
 export * from './schemas';
-export * from './types';
 export * from './util';
