+++ src/db/util/init.ts
@@ -8,7 +8,7 @@ import { codeBundle, contract, user } from '../schemas';
 
 import { getUser } from '../queries/user';
 import { getStoredPrivateKey } from './identity';
-import { UserDocument } from '@db/types';
+import { UserDocument } from 'types';
 
 function isLocalNode(rpcUrl: string): boolean {
   return !rpcUrl.includes('127.0.0.1');
