+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import '../styles/main.css';
-import { App } from '@ui/components';
+import { App } from 'ui/components';
 
 const root = document.getElementById('app-root');
