+++ src/types/index.ts
@@ -1,4 +1,5 @@
 // Copyright 2021 @paritytech/canvas-ui-v2 authors & contributors
 
+export * from './db';
 export * from './substrate';
 export * from './app';
